changed ints to floats and added fastenemy

+++ src/FastEnemy.java
@@ -16,7 +16,7 @@ public class FastEnemy extends GameObject{
 	}
 	
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 20, 20);
+		return new Rectangle((int )x,(int) y, 20, 20);
 	}
 
 	public void tick() {
@@ -32,7 +32,7 @@ public class FastEnemy extends GameObject{
 
 	public void render(Graphics g) {
 		g.setColor(Color.CYAN);
-		g.fillRect(x, y, 20, 20);
+		g.fillRect((int) x,(int) y, 20, 20);
 		
 	}
